% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-census.R
\name{acs_demographic_data}
\alias{acs_demographic_data}
\title{ACS Census demographic data}
\usage{
acs_demographic_data(acs_variables, state, tract_fips)
}
\description{
Import demographic data from the ACS. Import multiple tables at once and return in single dataframe.
}
\keyword{internal}
