% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-census.R
\name{acs_tracts_race}
\alias{acs_tracts_race}
\title{Create data set with racial population and percentage breakdowns by census tracts}
\usage{
acs_tracts_race(.data, races_to_use)
}
\arguments{
\item{.data}{Data set containing imported census demographic data created with \code{acs_demographic_data()}}

\item{races_to_use}{Races that we want to include in the data, as a vector of strings.}
}
\description{
Create data set with racial population and percentage breakdowns by census tracts
}
\keyword{internal}
