% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-viz-tables.R
\name{add_title_axis_labels}
\alias{add_title_axis_labels}
\title{Axis labels for Highcharter line charts}
\usage{
add_title_axis_labels(plt, plt_title, plt_x_label = "School Year", plt_y_label)
}
\description{
Axis labels for Highcharter line charts
}
