% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{assessment_scores_by_race}
\alias{assessment_scores_by_race}
\title{Create table with state-level assessment scores for the whole state and a single district}
\usage{
assessment_scores_by_race(state_assessment_data, org_level, nces_num)
}
\arguments{
\item{state_abb}{Two letter state abbreviation.}

\item{district_leaid}{The district's LEAID number, as an integer.}

\item{years}{The years from which you want to retrieve data. No data will be returned if data
is not available for the year.}

\item{grade}{Grades to import. Defaults to all grades (99)}
}
\value{
A single data set that contains assessment scores broken down by race. It includes rows
for the whole state and a single district that is specified.
}
\description{
Create table with state-level assessment scores for the whole state and a single district
}
