% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-census.R
\name{calculate_percentages}
\alias{calculate_percentages}
\title{Covert raw counts to percentages}
\usage{
calculate_percentages(.data, total_variable_string)
}
\description{
Take a data set of raw count numbers and calcualte percentages for groups.
}
\keyword{internal}
