% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shapefiles.R
\name{calculate_tracts_in_district}
\alias{calculate_tracts_in_district}
\title{Helper function that finds tracts in district}
\usage{
calculate_tracts_in_district(district_shapefile, state_census_tracts_shapefile)
}
\description{
Helper function that finds tracts in district
}
