% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shapefiles.R
\name{census_tracts_in_district}
\alias{census_tracts_in_district}
\title{Census tracts in district}
\usage{
census_tracts_in_district(state, district_shapefile)
}
\description{
Find the census tracts in the district and return a list containing the following elements:
(1) vector of tract IDs that are in the district, (2) shapefile of tracts in the district.
}
