% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{check_level}
\alias{check_level}
\title{Check 'level' parameter in functions}
\usage{
check_level(level)
}
\description{
For the educationdata API, the level must be either 'schools' or 'school-districts' for many calls.
This function checks to ensure that the level parameter is one of these two values and throws an error if not.
}
\keyword{internal}
