% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-census.R
\name{choropath_tracts_race}
\alias{choropath_tracts_race}
\title{Create choropath of census tracts and racial brakdowns}
\usage{
choropath_tracts_race(
  .data,
  race,
  tracts_in_district,
  district_shapefile,
  specific_race_population,
  school_directory
)
}
\description{
Create choropath of census tracts and racial brakdowns
}
\keyword{internal}
