% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-viz-tables.R
\name{cities_in_district}
\alias{cities_in_district}
\title{Table of cities}
\usage{
cities_in_district(district_cities)
}
\description{
Clean the table of cities in the district and convert it to a gt table
}
