% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{clean_enrollment_by_race}
\alias{clean_enrollment_by_race}
\title{Clean enrollment by race data}
\usage{
clean_enrollment_by_race(race_enrollment)
}
\arguments{
\item{race_enrollment}{Data frame with enrollment data by race (school or district), created with get_ccd_enrollment()}
}
\value{
A data frame with percentage race breakdowns by year.
}
\description{
Calculate percentage race breakdowns, change race names, reorder race
}
