% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{clean_numeric_assessments}
\alias{clean_numeric_assessments}
\title{Clean up numeric values of state assessments and calculate number who passed}
\usage{
clean_numeric_assessments(.data)
}
\description{
Clean up numeric values of state assessments and calculate number who passed
}
