% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata.R
\name{convert_to_na}
\alias{convert_to_na}
\title{Convert educationdata api missing values to NA}
\usage{
convert_to_na(.data, cols)
}
\description{
Convert educationdata api missing values to NA
}
\keyword{internal}
