% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{convert_to_sy}
\alias{convert_to_sy}
\title{Convert year as a single number to a school year}
\usage{
convert_to_sy(years_col)
}
\value{
A vector of school years.
}
\description{
For example, convert 2014 to 2014-2015. For all education data API years, the year listed is the
fall semester. Function also orders the school year column by the year column, so plots with the
school year column on the x axis are in the proper order
}
\keyword{internal}
