% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiles-educationdata.R
\name{crdc_years}
\alias{crdc_years}
\title{Get years CRDC is available}
\usage{
crdc_years(years)
}
\description{
CRDC data is not available every year. This function takes a sequence of years and returns
a vector of years that CRDC data is available
}
\keyword{internal}
