% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-viz-tables.R
\name{create_html_tooltip}
\alias{create_html_tooltip}
\title{Formatting for Highcharts tooltip}
\usage{
create_html_tooltip(group_color, y_text, total_number)
}
\description{
Formatting for Highcharts tooltip
}
