% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{data_years_available}
\alias{data_years_available}
\title{Years data is available for each Urban Inst. API data source}
\usage{
data_years_available(data_source)
}
\arguments{
\item{data_source}{The data source for the API. One of 'ccd', 'crdc', 'edfacts', or 'saipe}
}
\value{
Vector with the years (as integers) in which the data source is available.
}
\description{
Years data is available for each Urban Inst. API data source
}
