% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-census.R
\name{educational_attainment}
\alias{educational_attainment}
\title{Calculate educational attainment}
\usage{
educational_attainment(state, tract_fips, year)
}
\description{
Pull in educational attainment data, combine educational levels, and calcualte the percentage
of the populatin that has attained certain educational levels.
}
\keyword{internal}
