% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata.R
\name{get_all_educationdata}
\alias{get_all_educationdata}
\title{Get education data}
\usage{
get_all_educationdata(nces_number, years, data_sources)
}
\arguments{
\item{nces_number}{The school's NCES number, as a string.}

\item{years}{The years from which you want to retrieve data. No data will be returned if data
is not available for the year.}

\item{data_sources}{The data sources to retrieve. Can be one of the following:
enrollment race', 'enrollment lep', 'test participation', 'IB AP enrollment', 'state assessments'}
}
\value{
A list with each element containing a data frame with school data. The list is named,
which identifies the data contained in the list element.
}
\description{
Imports multiple pieces of data from the Urban Institute's educationdata API at once. Users can specify
the needed data sets and all data sets will be returned in a single list.
}
