% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{get_ccd_directory}
\alias{get_ccd_directory}
\title{Get directory (staff, address, etc.)}
\usage{
get_ccd_directory(org_level, nces_num, year)
}
\arguments{
\item{org_level}{The level to return the results. This is the 'level' parameter in the API.
One of either 'schools' or 'school-district'}

\item{nces_num}{The school's LEAID number or NCES number, as a string.}

\item{year}{The years from which you want to retrieve data. No data will be returned if data
is not available for the year.}
}
\value{
A data set with the district directory.
}
\description{
Get directory (staff, address, etc.)
}
