% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{get_ccd_enrollment}
\alias{get_ccd_enrollment}
\title{Get enrollment (staff, address, etc.)}
\usage{
get_ccd_enrollment(org_level, nces_num, years, grades)
}
\arguments{
\item{org_level}{The level to return the results. This is the 'level' parameter in the API.
One of either 'schools' or 'school-district'}

\item{nces_num}{The school's LEAID number or NCES number, as a string.}

\item{years}{The years from which you want to retrieve data. No data will be returned if data
is not available for the year.}

\item{grades}{The grades to import with API. Enrollment numbers will combine all grades into one number.}
}
\value{
A data set with the enrollment data.
}
\description{
Get enrollment (staff, address, etc.)
}
