% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiles-educationdata.R
\name{get_ccd_school_directory}
\alias{get_ccd_school_directory}
\title{Import school overview data}
\usage{
get_ccd_school_directory(nces_number, year)
}
\description{
Import data from the educationdata api that is an overview of the school.
Uses the "school ccd directory" endpoint.
}
\keyword{internal}
