% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiles-educationdata.R
\name{get_ccd_school_enrollment_race}
\alias{get_ccd_school_enrollment_race}
\title{Enrollment by race}
\usage{
get_ccd_school_enrollment_race(nces_number, years)
}
\description{
Import data from the educationdata api that is school enrollment by race.
Uses the "school ccd enrollment" endpoint
}
\keyword{internal}
