% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata.R
\name{get_crdc_school_enrollment_lep}
\alias{get_crdc_school_enrollment_lep}
\title{Enrollment by limit English proficiency (LEP)}
\usage{
get_crdc_school_enrollment_lep(nces_number, years)
}
\description{
Import data from the educationdata api that is school enrollment by LEP status
Uses the "school crdc enrollment lep sex" endpoint
}
\keyword{internal}
