% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiles-educationdata.R
\name{get_crdc_school_test_participation}
\alias{get_crdc_school_test_participation}
\title{ACT / SAT participation}
\usage{
get_crdc_school_test_participation(nces_number, years)
}
\description{
Returns the number of students taking the ACT/ SAT by denographic group
}
\keyword{internal}
