% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata.R
\name{get_crdc_school_test_participation_percentages}
\alias{get_crdc_school_test_participation_percentages}
\title{Percentage of ACT / SAT participation by race}
\usage{
get_crdc_school_test_participation_percentages(
  nces_number,
  years,
  crdc_enrollment_by_race
)
}
\arguments{
\item{crdc_enrollment_by_race}{A data set of  CRDC enrollment by race. Needed so we can calculate percentages
Created with get_crdc_topic_subtopics(nces_number, years, "enrollment", c("race", "sex")) \%>\% clean_crdc('enrollment_crdc')}
}
\description{
Returns the percentage of students taking the ACT/ SAT by race
}
\keyword{internal}
