% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{get_district_in_poverty}
\alias{get_district_in_poverty}
\title{Percentage of 5-17 year olds in poverty in district}
\usage{
get_district_in_poverty(leaid, year)
}
\arguments{
\item{leaid}{The school's LEAID number, as a string.}

\item{year}{The years from which you want to retrieve data. No data will be returned if data
is not available for the year.}
}
\value{
A data set with the percentage of 5-17 year old sin poverty by year.
}
\description{
Percentage of 5-17 year olds in poverty in district
}
