% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata.R
\name{get_eddata_topic_subtopic}
\alias{get_eddata_topic_subtopic}
\title{Base function to import data from the educationdata api}
\usage{
get_eddata_topic_subtopic(
  nces_number,
  years,
  source,
  topic,
  subtopic = NULL,
  grades = FALSE
)
}
\description{
Base function to import data from the educationdata api
}
\keyword{internal}
