% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{get_edfacts_state_assessments}
\alias{get_edfacts_state_assessments}
\title{Get state assessment data for a district}
\usage{
get_edfacts_state_assessments(org_level, fips_code, years, grade = 99)
}
\arguments{
\item{org_level}{The level to return the results. This is the 'level' parameter in the API.
One of either 'schools' or 'school-district'}

\item{fips_code}{State FIPS code, as integer.}

\item{years}{The years from which you want to retrieve data. No data will be returned if data
is not available for the year.}

\item{grade}{Grades to import. Defaults to all grades (99)}
}
\value{
A list with each element containing a data frame with school data. The list is named,
which identifies the data contained in the list element.
}
\description{
Imports district data from Ed Facts using the Urban Institute's API. The data contains state assessment information.
}
