% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{get_schools_in_district}
\alias{get_schools_in_district}
\title{School directory for schools in district and grade}
\usage{
get_schools_in_district(leaid, year, grades = 99)
}
\arguments{
\item{leaid}{The school's LEAID number, as a string.}

\item{year}{The years from which you want to retrieve data. No data will be returned if data
is not available for the year.}

\item{grades}{Grades to import. Defaults to all grades (99)}
}
\value{
A data set with directory information for all schools in the district and within grades.
}
\description{
School directory for schools in district and grade
}
