% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{get_state_school_numbers}
\alias{get_state_school_numbers}
\title{Create dataset with all district / school names and their LEAID / NCES numbers}
\usage{
get_state_school_numbers(org_level, state_abb, year)
}
\description{
Used so we can locate leaid numbers from district names
}
