% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-viz-tables.R
\name{hc_plot_grouped_bar}
\alias{hc_plot_grouped_bar}
\title{Highcharts grouped bar chart}
\usage{
hc_plot_grouped_bar(.data, x_col, y_col, group_col, y_title)
}
\description{
Highcharts grouped bar chart
}
