% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-viz-tables.R
\name{hc_plot_grouped_line}
\alias{hc_plot_grouped_line}
\title{Highcharts grouped line chart with percentages}
\usage{
hc_plot_grouped_line(
  .data,
  x_col,
  y_col,
  group_col,
  plt_title,
  x_var_title,
  y_var_title,
  y_percentage = TRUE
)
}
\description{
Highcharts grouped line chart with percentages
}
