% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{identify_school_grades}
\alias{identify_school_grades}
\title{Identify if school contains a grade}
\usage{
identify_school_grades(.data, vector_of_grades)
}
\description{
NCES data reports grades for schools, but reports it as the lowest and highest grades.
This function takes a vector of grades as input and then returns a boolean of whether the school
contains any of the grades based on the lowest and highest grades.
}
