% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit-dashboard-with-params.R
\name{knit_dashboard}
\alias{knit_dashboard}
\title{Create district flexdashboard for a given school district}
\usage{
knit_dashboard(
  output_filename,
  district_leaid_number,
  years,
  state_abb,
  grades
)
}
\description{
Create district flexdashboard for a given school district
}
