% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-viz-tables.R
\name{leaflet_census_tracts}
\alias{leaflet_census_tracts}
\title{Plot data by census tract using Leaflet}
\usage{
leaflet_census_tracts(
  district_tracts_shapefiles,
  census_data_by_tract,
  district_shapefile_for_centroid,
  school_location,
  tool_tip_labels
)
}
\arguments{
\item{district_tracts_shapefiles}{A shape file  with the district tracts, must contain a 'GEOID' column.}

\item{census_data_by_tract}{Data frame with data by census tract taht we want to plot. Must contain
a 'GEOID' column because this column is merged with the shapefile. There should only be one row per GEOID.}

\item{district_shapefile_for_centroid}{A shapefile for the school district, so we can compute the district's
centriod. This centroid is used to center the map,.}

\item{school_location}{A data frame containing the latitude and longitude of the school/s. The columns
for these two data points should be called 'latitude' and 'longitude'.}

\item{tool_tip_labels}{Tool tips to use, as HTML.
Example: \verb{glue::glue("<strong>\{census_data_by_tract$NAME\}</strong> |> lapply(htmltools::HTML)}}
}
\description{
Leaflet plot showing census tracts within a district a fill color for the tracts based on data included
the parameters.
}
