% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-viz-tables.R
\name{leaflet_district_schools}
\alias{leaflet_district_schools}
\title{District boundary and school plot}
\usage{
leaflet_district_schools(district_shapefile, school_information)
}
\description{
Leaflet plot showing district bounadries as a shapefile and a point for the school.
}
