% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{pretty_percent_cols}
\alias{pretty_percent_cols}
\title{Create percentage columns for plotting}
\usage{
pretty_percent_cols(.data, perc_col)
}
\description{
Create percentage columns for plotting
}
\keyword{internal}
