% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{relevel_race}
\alias{relevel_race}
\title{Relevel race factors}
\usage{
relevel_race(race_col)
}
\description{
Releve race factors in order of Black, Hispanic / Latinx, White, and Two or more at the end
}
\keyword{internal}
