% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{school_directory}
\alias{school_directory}
\title{Directory of school information.}
\format{
A data frame with 101688 rows and 8 variables:
}
\source{
\url{https://educationdata.urban.org/documentation/schools.html}
}
\usage{
school_directory
}
\description{
A dataset containing school directory information such as school name and number, district name and number,
state, and grades. Data comes from the Common Core of Data, retrieved from the Urban Institute's
Education Data Portal.
}
\keyword{datasets}
