% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{school_nces_numbers}
\alias{school_nces_numbers}
\title{Get a data frame with all schools from a state along with their key information}
\usage{
school_nces_numbers(state_abbreviation)
}
\arguments{
\item{state_abbreviation}{The two letter state abbbreviation on the state that
you want to show all schools for.}
}
\value{
A data frame showing all schools in a state, along with key identifying information.
}
\description{
Returns a data frame of all schools in a state, along with their nces number, leaid,
district, and address. This information is useful to find a school's nces number for
use in \code{school_report()}.
}
