% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit-dashboard-with-params.R
\name{school_report}
\alias{school_report}
\title{Create district flexdashboard for a given school district}
\usage{
school_report(output_filename, org_level, nces_num, years, state_abb, grades)
}
\description{
Create district flexdashboard for a given school district
}
