% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shapefiles.R
\name{shapefiles_district}
\alias{shapefiles_district}
\title{Import district shapefile}
\usage{
shapefiles_district(state, leaid_number, year)
}
\description{
Shapefiles for school boundaries do not exist. The closest we have are district boundary shapefiles.
}
\keyword{internal}
