% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{sr_add_leading_zeroes}
\alias{sr_add_leading_zeroes}
\title{Helper function to add leading zeroes to school and district numbers}
\usage{
sr_add_leading_zeroes(.data)
}
\arguments{
\item{.data}{Data frame containing columns titled \code{ncessch} and \code{leaid}
that we want to convert to character}
}
\description{
School and district numbers started receiveing an additional leading zero
in 2020. This function places all numbers in the same format by ensuring all
numbers are of the same length. Leading zeroes are added when numbers are not of the proper length
}
\keyword{internal}
