% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{sr_aggregate_assessment}
\alias{sr_aggregate_assessment}
\title{Aggregate state assessment data}
\usage{
sr_aggregate_assessment(.data, grouping_vars)
}
\description{
Summarize aggregate percentages of state assessment data by using the total number of takers
and total numebr of passers
}
\keyword{internal}
