% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{sr_assessment_scores_by_race}
\alias{sr_assessment_scores_by_race}
\title{Create a table with state-level assessment scores for the whole state and a single district}
\usage{
sr_assessment_scores_by_race(state_assessment_data, org_level, nces_num)
}
\arguments{
\item{state_assessment_data}{State assessment data set.}

\item{org_level}{The level to return the results. This is the 'level' parameter in the API.
One of either 'schools' or 'school-district'}

\item{nces_num}{The district's LEAID number or school's NCES number, as a string.}
}
\value{
A single data set that contains assessment scores broken down by race. It includes rows
for the whole state and a single district that is specified.
}
\description{
Returns a single data set with school / district and state assessment scores, which allows
for comparisons between the school / district and state.
}
