% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shapefiles.R
\name{sr_calculate_tracts_in_district}
\alias{sr_calculate_tracts_in_district}
\title{Helper function that finds tracts in district}
\usage{
sr_calculate_tracts_in_district(
  district_shapefile,
  state_census_tracts_shapefile
)
}
\description{
Helper function that finds tracts in district
}
\keyword{internal}
