% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{sr_ccd_directory}
\alias{sr_ccd_directory}
\title{Get CCD directory of a school or district.}
\usage{
sr_ccd_directory(org_level, nces_num, year)
}
\arguments{
\item{org_level}{The level to return the results. This is the 'level' parameter in the API.
One of either 'schools' or 'school-district'}

\item{nces_num}{The school's LEAID number or NCES number, as a string. Can use multiple
schools or districts by combining all numbers in a vector.}

\item{year}{The years from which you want to retrieve data. No data will be returned if data
is not available for the year.}
}
\value{
A data set with the district directory.
}
\description{
Returns the directory of a school or district for a single year, from the Common
Core of Data data set. In the Education Data API, this returns the "directory" topic from the "ccd" source.
Can return the information from multiple schools or districts by supplying a vector
of numbers to \code{nces_num}.
}
