% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{sr_ccd_enrollment}
\alias{sr_ccd_enrollment}
\title{Get enrollment data of the school or district.}
\usage{
sr_ccd_enrollment(org_level, nces_num, years, grades)
}
\arguments{
\item{org_level}{The level to return the results. This is the 'level' parameter in the API.
One of either 'schools' or 'school-district'}

\item{nces_num}{The district's LEAID number or school's NCES number, as a string.}

\item{years}{The years from which you want to retrieve data. No data will be returned if data
is not available for the year.}

\item{grades}{The grades to import. Enrollment numbers will combine all grades into one number.}
}
\value{
A data frame with the enrollment data.
}
\description{
Gets enrollment data of the school or district, by race. Data comes from the CCD
and is imported via the Education Data Portal API.
}
