% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shapefiles.R
\name{sr_census_tracts_in_district}
\alias{sr_census_tracts_in_district}
\title{Identify census tracts in a school district}
\usage{
sr_census_tracts_in_district(district_shapefile, state_abb)
}
\arguments{
\item{district_shapefile}{The shapefile of the district. Can use \code{tigris::school_districts()}
to get the shapefile.}

\item{state_abb}{The two letter state abbbreviation of the state in which
the district is located.}
}
\description{
Find the census tracts in the district and return a list containing the following elements:
(1) vector of tract IDs that are in the district, (2) shapefile of tracts in the district.
}
\examples{
\dontrun{
state_abb <- 'AR'

district_shapefile <- tigris::school_districts(state = state_abb, cb = FALSE) |>
   dplyr::filter(GEOID == "0503060")

sr_census_tracts_in_district(district_shapefile, state_abb)
}

}
