% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{sr_check_required_cols}
\alias{sr_check_required_cols}
\title{Check Required Columns}
\usage{
sr_check_required_cols(.data, required_cols)
}
\arguments{
\item{.data}{The data frame to check.}

\item{required_cols}{A vector of required column names.}
}
\value{
A logical value indicating if all the required columns are present.
}
\description{
This function checks if all the required columns are present in a data frame.
}
\keyword{internal}
