% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shapefiles.R
\name{sr_cities_in_district}
\alias{sr_cities_in_district}
\title{Identify cities in the district}
\usage{
sr_cities_in_district(district_shapefile, state_abb)
}
\arguments{
\item{district_shapefile}{The shapefile of the district. Can use \code{tigris::school_districts()}
to get the shapefile.}

\item{state_abb}{The two letter state abbbreviation on the state that
you want to show all schools for.}
}
\value{
A data frame with all the cities within a district's boundaries
and each city's population.
}
\description{
Get the cities in a school district and return a dataframe with the cities and their populations.
}
\examples{
\dontrun{
district_shapefile <- tigris::school_districts(state = "AR", cb = FALSE) |>
   dplyr::filter(GEOID == "0503060")

sr_table_cities_in_district(district_shapefile, "AR")
}

}
