% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{sr_clean_enrollment_by_race}
\alias{sr_clean_enrollment_by_race}
\title{Clean enrollment by race data}
\usage{
sr_clean_enrollment_by_race(.data)
}
\arguments{
\item{.data}{Data frame with enrollment data by race (school or district), created with \code{get_ccd_enrollment()}.}
}
\value{
A data frame with percentage race breakdowns by year.
}
\description{
Calculate percentage breakdowns by race and clean racial category names.
}
