% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{sr_clean_numeric_assessments}
\alias{sr_clean_numeric_assessments}
\title{Clean up numeric values of state assessments and calculate number who passed}
\usage{
sr_clean_numeric_assessments(.data)
}
\arguments{
\item{.data}{The state assessment data to clean.}
}
\description{
Clean up numeric values of state assessments and calculate number who passed
}
\keyword{internal}
