% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-census.R
\name{sr_demographic_data}
\alias{sr_demographic_data}
\title{Import ACS census data for a state and tracts}
\usage{
sr_demographic_data(acs_variables, state_abb, tract_fips)
}
\arguments{
\item{acs_variables}{The variable names from teh ACS that we want to pull.}

\item{state_abb}{The two letter state abbreviation on the state that
you want to show all schools for.}

\item{tract_fips}{The fips code of the census tracts that we want to get data for.}
}
\description{
Imports census tract and state-level demographic data from the ACS.
Then combines the census tract and state-level information into one data set.
This is a helper function for creating the Rmarkdown file and has little application
outside this file.
}
