% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{sr_district_in_poverty}
\alias{sr_district_in_poverty}
\title{Percentage of 5-17 year olds in poverty in district}
\usage{
sr_district_in_poverty(leaid, years)
}
\arguments{
\item{leaid}{The district's LEAID number, as a string.}

\item{years}{The years from which you want to retrieve data. No data will be returned if data
is not available for the year. Use a vector of years to return data from
multiple years.}
}
\value{
A data set with the percentage of 5-17 year olds in poverty by year.
}
\description{
Percentage of 5-17 year olds in poverty in district
}
