% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit-dashboard-with-params.R
\name{sr_download_template_rmd}
\alias{sr_download_template_rmd}
\title{Download the template Rmarkdown file}
\usage{
sr_download_template_rmd(destination_path)
}
\arguments{
\item{destination_path}{The path where the Rmarkdown file should be saved.}
}
\description{
This function allows users to download and save the template Rmarkdown file
that is used to build the default report. Users can then adjust the file
to suit their needs.
}
