% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{sr_identify_school_grades}
\alias{sr_identify_school_grades}
\title{Identify if school contains a grade}
\usage{
sr_identify_school_grades(.data, vector_of_grades)
}
\arguments{
\item{.data}{Dataset of CCD data, created with \code{\link[=sr_ccd_directory]{sr_ccd_directory()}}.}

\item{vector_of_grades}{Vector of integers representing the grades of which
we want to see if the school has any of the grades in the vector.}
}
\description{
NCES data reports grades for schools, but reports it as the lowest and highest grades.
This function takes a vector of grades as input and then returns a boolean of whether the school
contains any of the grades based on the lowest and highest grades.
}
