% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-viz-tables.R
\name{sr_plot_add_title_axis_labels}
\alias{sr_plot_add_title_axis_labels}
\title{Axis labels for Highcharter line charts}
\usage{
sr_plot_add_title_axis_labels(
  plt,
  plt_title,
  plt_x_label = "School Year",
  plt_y_label
)
}
\description{
Axis labels for Highcharter line charts
}
\keyword{internal}
