% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-viz-tables.R
\name{sr_plot_choropleth_tracts}
\alias{sr_plot_choropleth_tracts}
\title{Choropleth maps by census tract using leaflet}
\usage{
sr_plot_choropleth_tracts(
  census_data_by_tract,
  tract_shapefile,
  school_location,
  tool_tip_labels
)
}
\arguments{
\item{census_data_by_tract}{Data frame with data by census tract that we want to plot. Must contain
a 'GEOID' column because this column is merged with the census tract shapefile (\code{tract_shapefile}). There should only be one row per GEOID.}

\item{tract_shapefile}{Shapefiles for the same census tracts as \code{census_data_by_tract}.
Can be imported with \verb{[sr_census_tracts_in_district()]}. Can also use \verb{[tigris::tracts()]}. Can also use}

\item{school_location}{A data frame containing the latitude and longitude of the school/s. The columns
for these two data points should be called 'latitude' and 'longitude'. This information
can be retreived from CCD school directories, which can be imported with \code{\link[=sr_ccd_directory]{sr_ccd_directory()}}.}

\item{tool_tip_labels}{Tool tips to use, as HTML. Example,  which only shows the census tract name:
\verb{glue::glue("<strong>\{census_data_by_tract$NAME\}</strong> |> lapply(htmltools::HTML)}}
}
\value{
A leaflet choropleth map of census tracts.
}
\description{
Leaflet plot showing census tracts within a district and a fill color for the tracts based on data included
in the parameters.
}
