% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-viz-tables.R
\name{sr_plot_create_tooltip}
\alias{sr_plot_create_tooltip}
\title{Create Highcharts tooltip for plots}
\usage{
sr_plot_create_tooltip(plt, group_color, y_text, total_number)
}
\arguments{
\item{plt}{Plot in which to add tool tip.}

\item{group_color}{Column name, as a string, that represents a grouping variable.}

\item{y_text}{The column name, as a string, of the y-axis variable.}

\item{total_number}{The column name, as a string of a column containing a number
we want to add to the tool tip.}
}
\description{
Creates tool tips in a style that works for \code{schoolreportr} plots. May not work for other plots.
Works best for grouped line charts.
}
