% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-viz-tables.R
\name{sr_plot_grouped_line}
\alias{sr_plot_grouped_line}
\title{Grouped line chart}
\usage{
sr_plot_grouped_line(
  .data,
  x_col,
  y_col,
  group_col,
  plt_title = NULL,
  x_var_title = NULL,
  y_var_title = NULL,
  y_percentage = FALSE
)
}
\arguments{
\item{.data}{The data frame to plot.}

\item{x_col}{The column name, as a string, of the plot's x-axis. Generally, this
will be a time variable such as years.}

\item{y_col}{The column name, as a string, of the y-axis. If this column is a percentage,
it should be as a whole number (78.8) and not a decimal (.788).}

\item{group_col}{The column name, as a string, of the grouping column. The plot
will produce different lines for each group in this column.}

\item{plt_title}{The plot title, as a string.}

\item{x_var_title}{The x-axis title, as a string.}

\item{y_var_title}{The y-axis title, as a string.}

\item{y_percentage}{Boolean, signifying whether the y-axis is a percentage.
Defaults to \code{FALSE}.}
}
\value{
A highcharts line plot
}
\description{
Produces a highcharts grouped line chart. Since the data, axis, and grouping
variables are parameters in the function, this function will work with any data.
}
