% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-viz-tables.R
\name{sr_plot_percentage}
\alias{sr_plot_percentage}
\title{Create percentages on axis labels}
\usage{
sr_plot_percentage(plt, y_var_title)
}
\description{
Create percentages on axis labels
}
\keyword{internal}
