% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-viz-tables.R
\name{sr_plot_total_enrollment_by_year}
\alias{sr_plot_total_enrollment_by_year}
\title{Plot total enrollment by year}
\usage{
sr_plot_total_enrollment_by_year(.data, school_names, grade_span)
}
\arguments{
\item{.data}{Data frame containing enrollment data. This data frames comes from
\code{sr_ccd_enrollment} and is then filtered to only contain total enrollment.}

\item{school_names}{A string containing the school name(s)}

\item{grade_span}{A string containing which grades are included (example: \code{"10-12"}).
This information is used in the plot title.}
}
\value{
A highcharts line plot.
}
\description{
Create a highcharts line plot that shows total enrollment in all schools by year.
This function has little use outside the Rmarkdown file as it relies on the
specific data in the Rmarkdown file.
}
\examples{
\dontrun{
total_enrollment <- sr_ccd_enrollment("schools", "050306000073", 2017:2020, 10:12) |>
   dplyr::filter(race == 'Total')

sr_plot_total_enrollment_by_year(total_enrollment, "Bentonville High School", "10-12")
}

}
