% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{sr_rename_reorder_race_education}
\alias{sr_rename_reorder_race_education}
\title{Rename and reorder race categories}
\usage{
sr_rename_reorder_race_education(race_col)
}
\arguments{
\item{race_col}{A vector containing the race data}
}
\value{
A factor vector with race names changes and order changes.
}
\description{
Rename Black to 'Black / African-American' and Hispanic to 'Hispanic / Latinx'. Also reorder so that
'Black / African-American' and 'Hispanic / Latinx' appear first and second.
}
\keyword{internal}
