% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-census.R
\name{sr_rmd_census_tracts_race}
\alias{sr_rmd_census_tracts_race}
\title{Create data set with racial population and percentage breakdowns by census tracts}
\usage{
sr_rmd_census_tracts_race(.data, races_to_use)
}
\arguments{
\item{.data}{Data set containing imported census demographic data created with \code{\link[=sr_demographic_data]{sr_demographic_data()}}}

\item{races_to_use}{Races that we want to include in the data, as a vector of strings.}
}
\description{
Cleans ACS census data with race variables by putting it into a format where it
can be plotted with a choropleth map in highcharts.
}
\details{
Helper function that is used to create Rmarkdown file
}
