% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-viz-tables.R
\name{sr_rmd_plot_choropleth_tracts_race}
\alias{sr_rmd_plot_choropleth_tracts_race}
\title{Choropleth map of census tracts and racial brakdowns}
\usage{
sr_rmd_plot_choropleth_tracts_race(.data, race, ...)
}
\arguments{
\item{.data}{Data set of racial totals and percentages by census tract.
Data set created with \code{sr_rmd_census_tracts_race}. This does not contain
shapefile data.}

\item{race}{String representing racial group we want to plot. This will come from
the \code{variable} column in \code{.data}.}

\item{...}{Parameters for \code{sr_plot_choropleth_tracts}.}
}
\value{
A leaflet choropleth map of census tracts, with the percentage of
people in a given census tract belonging to a given racial identity.
}
\description{
Helper function that plots a choropleth map of the percentage of people of a given
race living in each census tract within a district. Uses \code{sr_plot_choropleth_tracts}
to create the plot, but creates and adds the tool tip and filters the data
for the racial category that we want to plot.
}
