% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit-dashboard-with-params.R
\name{sr_school_report}
\alias{sr_school_report}
\title{Create flexdashboard for a given school or list of schools}
\usage{
sr_school_report(output_filename, nces_num, years, state_abb, grades)
}
\arguments{
\item{output_filename}{The filename and path to output the HTML file of the school report.
The path is relative to your R root folder in your current session. It should end in '.html'.}

\item{nces_num}{The NCES number(s) of the school(s) that we want to generate
reports for. The numbers should be as strings. Use a vetor of numbers if
to aggregate the data of multiple schools within a single district.
Use \code{\link[=sr_school_nces_numbers]{sr_school_nces_numbers()}} to get a data frame of NCES numbers for each school in a state.}

\item{years}{The years, as a vector of integers, we want to include in the data.}

\item{state_abb}{The two letter state abbreviation where the school(s) are located.}

\item{grades}{A vector of integers representing the grades we want included in the report.
If the school does not contain the grades requested, these grades will be ignored.
Use \code{\link[=sr_school_nces_numbers]{sr_school_nces_numbers()}} to see which grades each school in a state serves.
Use -1 for Pre-K and 0 for K.}
}
\description{
Use the automated template to create a report for a school that provides background
information about the school or list of schools. If more than one school is specified,
all the data from the data from the schools is aggregated. But, all school must
be in the same district.
}
