% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{sr_state_fips_code}
\alias{sr_state_fips_code}
\title{Get state FIPS codes from state abbreviations}
\usage{
sr_state_fips_code(state_abb)
}
\arguments{
\item{state_abb}{Two letter state abbreviation of the state that you want to
retrieve the fips code.}
}
\value{
A single integer representing the fips code of the specified state.
}
\description{
Pulling state data from APIs often requires knowing teh state's fips code, which
is an integer. This function allows you to retreive a single state's fips code
by entering the state abbreviation.
}
