% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-viz-tables.R
\name{sr_viz_district_boundary}
\alias{sr_viz_district_boundary}
\title{District boundary and school plot}
\usage{
sr_viz_district_boundary(district_shapefile, school_information)
}
\arguments{
\item{district_shapefile}{The shapefile of the district. Can use \code{tigris::school_districts()}
to get the shapefile.}

\item{school_information}{A data frame containing information about the school, including
latitdue and longitude location data. Can get this information from \code{sr_ccd_directory()}}
}
\value{
A leaflet plot with the district boundaries and the school locations.
}
\description{
Leaflet plot showing district boundaries as a shapefile and a point for schools.
The tooltip over teh school contains school infomration.
}
\examples{
\dontrun{
district_shapefile <- tigris::school_districts(state = "AR", cb = FALSE) |>
   dplyr::filter(GEOID == "0503060")

school_directory <- sr_ccd_directory('schools', "050306000073", 2021)

sr_viz_district_boundary(district_shapefile, school_directory)
}


}
