% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-educationdata-district.R
\name{sr_years_data_available}
\alias{sr_years_data_available}
\title{Years data is available for each Urban Inst. API data source}
\usage{
sr_years_data_available(data_source)
}
\arguments{
\item{data_source}{The data source for the API. One of 'ccd', 'crdc', 'edfacts', or 'saipe}
}
\value{
Vector with the years (as integers) in which the data source is available.
}
\description{
Find the years each data source is available for data from the Urban Institute's
Education data explorer API.
}
