% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{state_fips_code}
\alias{state_fips_code}
\title{Get state FIPS codes from state abbreviations}
\usage{
state_fips_code(state_abb)
}
\description{
Get state FIPS codes from state abbreviations
}
\keyword{internal}
