% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shapefiles.R
\name{table_cities_in_single_district}
\alias{table_cities_in_single_district}
\title{Identify cities in the district}
\usage{
table_cities_in_single_district(district_shapefile, state)
}
\description{
Get the cities in a school district and return a dataframe with the cities and their population.
}
